+++ iowmongotools/app.py
@@ -7,6 +7,8 @@ from abc import ABC, abstractmethod
 from argparse import ArgumentParser, ArgumentDefaultsHelpFormatter
 import subprocess
 import yaml
+import re
+from collections import OrderedDict
 
 logging.basicConfig(format='%(asctime)s %(levelname)s: %(message)s', datefmt='%Y-%m-%d %H:%M:%S', level=logging.INFO)
 logger = logging.getLogger(__name__)
@@ -26,6 +28,102 @@ def run_ext_command(args, title=''):
     return proc.wait()
 
 
+def human_to_seconds(string):
+    """Convert internal string like 1M, 1Y3M, 3W to seconds.
+
+    :type string: str
+    :param string: Interval string like 1M, 1W, 1M3W4h2s...
+        (s => seconds,
+         m => minutes,
+         h => hours,
+         D => days,
+         W => weeks,
+         M => months,
+         Y => Years).
+
+    :rtype: int
+    :return: The conversion in seconds of string.
+    """
+    interval_dict = OrderedDict([("Y", 365 * 86400),  # 1 year
+                                 ("M", 30 * 86400),  # 1 month
+                                 ("W", 7 * 86400),  # 1 week
+                                 ("D", 86400),  # 1 day
+                                 ("h", 3600),  # 1 hour
+                                 ("m", 60),  # 1 minute
+                                 ("s", 1)])  # 1 second
+    interval_exc = "Bad interval format for {0}".format(string)
+
+    interval_regex = re.compile(
+        "^(?P<value>[0-9]+)(?P<unit>[{0}])".format("".join(interval_dict.keys())))
+    seconds = 0
+
+    while string:
+        match = interval_regex.match(string)
+        if match:
+            value, unit = int(match.group("value")), match.group("unit")
+            if int(value) and unit in interval_dict:
+                seconds += value * interval_dict[unit]
+                string = string[match.end():]
+            else:
+                raise Exception(interval_exc)
+        else:
+            raise Exception(interval_exc)
+    return seconds
+
+
+def deep_merge(a, b, path=None):
+    if path is None:
+        path = []
+    for key in b:
+        if key in a:
+            if isinstance(a[key], dict) and isinstance(b[key], dict):
+                deep_merge(a[key], b[key], path + [str(key)])
+            elif a[key] == b[key]:
+                pass  # same leaf value
+            else:
+                raise Exception('Conflict at %s' % '.'.join(path + [str(key)]))
+        else:
+            a[key] = b[key]
+    return a
+
+
+class DictConfigurator(logging.config.DictConfigurator):
+    def configure_logger(self, name, config, incremental=False):
+        """Configure a non-root logger from a dictionary. Extended with factory class."""
+        factory_class = config.get('()', None)
+        _logger_class_buf = logging.getLoggerClass()
+        if factory_class:
+            logging.setLoggerClass(getattr(sys.modules[__name__], factory_class))
+        logger = logging.getLogger(name)
+        logging.setLoggerClass(_logger_class_buf)
+        self.common_logger_config(logger, config, incremental)
+        propagate = config.get('propagate', None)
+        if propagate is not None:
+            logger.propagate = propagate
+
+
+logging.config.dictConfigClass = DictConfigurator
+
+
+class ConstantExtraLogger(logging.Logger):
+    """Extended with constant 'extra'."""
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.extra = dict()
+
+    def makeRecord(self, name, level, fn, lno, msg, args, exc_info,
+                   func=None, extra=None, sinfo=None):
+        rv = logging.LogRecord(name, level, fn, lno, msg, args, exc_info, func, sinfo)
+        extra = extra or self.extra
+        if extra is not None:
+            for key in extra:
+                if (key in ["message", "asctime"]) or (key in rv.__dict__):
+                    raise KeyError("Attempt to overwrite %r in LogRecord" % key)
+                rv.__dict__[key] = extra[key]
+        return rv
+
+
 class Settings(object):
     """ Container with settings loaded from defaults which may be overwritten in config file or cmd-arguments"""
 
@@ -34,8 +132,10 @@ class Settings(object):
             logger.debug('Loading %s default settings', len(defaults))
             self.description = {}
             self._load_defaults(defaults)
+        self.tmp = dict()
         self.load()
         delattr(self, 'description')
+        delattr(self, 'tmp')
 
     def _load_defaults(self, defaults):
         for key, content in defaults.items():  # content is tuple of value and description
@@ -81,13 +181,7 @@ class SettingsCli(Settings):
         config = self.read_config_file(self.config_file)
         if config:
             self.__dict__.update(config)
-        parser.add_argument("--{}".format('cluster_config'), default=getattr(self, 'cluster_config', None),
-                            help='Path to yaml file containing description of clusters')
-        parser.parse_known_args(args=[a for a in argv if a not in ['-h', '--help']], namespace=self)
-        cluster_config = self.read_config_file(self.cluster_config)
-        if cluster_config:
-            parser.add_argument("--{}".format('clusters'), default=[c for c in cluster_config.keys()],
-                                help='Cluster names to be processed', nargs='*')
+        exclusions = self.extra_run(parser, argv)
         for key, value in self.__dict__.items():
             if not self.description.get(key):
                 continue
@@ -99,15 +193,50 @@ class SettingsCli(Settings):
                                     help=self.description.get(key))
             elif isinstance(value, (list, tuple, set, frozenset)):
                 parser.add_argument("--{}".format(key), default=value, nargs='*', help=self.description.get(key))
-            elif key not in ('config_file', 'description', 'cluster_config'):
+            elif key not in ('config_file', 'description', *exclusions):
                 parser.add_argument("--{}".format(key), default=value, help=self.description.get(key))
         parser.parse_args(argv, self)
-        if cluster_config:
-            self.cluster_config = cluster_config
+        self.cleanup()
+        if not getattr(self, 'config_file'):
+            delattr(self, 'config_file')
+
+    def extra_run(self, parser, argv):
+        return ()
+
+    def cleanup(self):
+        pass
+
+
+class SettingCliCluster(SettingsCli):
+
+    def extra_run(self, parser, argv):
+        return self.add_clusters(parser, argv)
+
+    def cleanup(self):
+        if self.tmp['cluster_config']:
+            self.cluster_config = self.tmp['cluster_config']
             self.clusters = set(self.clusters)
-        for attr in ('config_file', 'cluster_config'):
-            if not getattr(self, attr):
-                delattr(self, attr)
+        if not getattr(self, 'cluster_config'):
+            delattr(self, 'cluster_config')
+
+    def add_clusters(self, parser, argv):
+        parser.add_argument("--{}".format('cluster_config'), default=getattr(self, 'cluster_config', None),
+                            help='Path to yaml file containing description of clusters')
+        parser.parse_known_args(args=[a for a in argv if a not in ['-h', '--help']], namespace=self)
+        self.tmp['cluster_config'] = self.read_config_file(self.cluster_config)
+        if self.tmp['cluster_config']:
+            parser.add_argument("--{}".format('clusters'), default=[c for c in self.tmp['cluster_config'].keys()],
+                                help='Cluster names to be processed', nargs='*')
+        return 'cluster_config',
+
+
+class SettingCliUploader(SettingCliCluster):
+
+    def extra_run(self, parser, argv):
+        if getattr(self, 'upload'):
+            parser.add_argument("--{}".format('providers'), default=[p for p in self.upload.keys()],
+                            help='List of providers of segments for processing', nargs='*')
+        return self.add_clusters(parser, argv)
 
 
 class Command(object):
@@ -143,8 +272,8 @@ class Invoker(object):
             for cmd in command:
                 self.add(cmd)
             return
-        if not issubclass(command.__class__, Command):
-            raise TypeError("Instanses of '%s' class are allowed only" % Command.__name__)
+        if not isinstance(command, Command):
+            raise TypeError("Instances of '%s' class are allowed only" % Command.__name__)
         self.registry.append(command)
 
     def execute(self, force=False):